Move utils to common_utils folder
