+++ framework/test/jest/unit/specs/modules/chain/dpos/round_delegates.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const BigNum = require('@liskhq/bignum');
-const { randomInt } = require('../../../../utils');
+const { randomInt } = require('../../../../../common_utils');
 const delegatePublicKeys = require('./delegate_publickeys.json');
 
 const delegateAccounts = delegatePublicKeys.map((pk, index) => {
